+++ source/OptionGrid.cpp
@@ -68,7 +68,7 @@ namespace QTWC {
         QQmlProperty::write(mOptions[option]->rootObject(), "checked", newState);
     }
 
-    void OptionGrid::setFontSize(int fontSize) {
+    void OptionGrid::setFontSize(double fontSize) {
         for (auto& mOption : mOptions) {
             QQmlProperty::write(mOption.second->rootObject(), "fontSize", fontSize);
         }
